backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Legent AI Backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy",
    "asyncpg",
    "alembic",
    "pydantic-settings",
]
